Get range from results of split and TestFunctions in SWE
We'll want to reuse the existing shallow water code for the plume model
but that model has a larger state space. This change makes the SWE demo
always pick out the first and second items of the tuple from either
`split` or `TestFunctions`, so if we add other variables like salinity
or temperature it'll just ignore them.