Allow for alternate field variables
Currently all our solvers assume that the fields being solved for are
the same as the conserved fields. For example for shallow water we have
to use thickness + momentum. This patch makes it so we can instead work
with e.g. thickness + velocity as long as we provide a mapping to the
conserved fields.