Get part of return value of split and TestFunctions
We'll want to reuse the existing shallow water / thin film code for the
plume model. But the plume model has a larger state space. This change
makes the SWE and thin film models always pick out just the first and
second items of the return value of `split` and `TestFunctions`. That
way we can easily tack on other variables like salinity or temperature.